pyproject.toml
[tool.poetry]
name = "san2patch"
version = "0.1.0"
description = ""
authors = ["san2patch.sec@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.3.14"
langgraph = "^0.2.64"
langchain-anthropic = "^0.3.3"
langchain-google-genai = "^2.0.9"
langchain-openai = "^0.3.1"
click = "^8.1.8"
psutil = "^6.1.1"
aim = "^3.27.0"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
coloredlogs = "^15.0.1"
peewee = "^3.17.8"
pyyaml = "^6.0.2"
oauth2client = "^4.1.3"
apiclient = "^1.0.4"
google-api-python-client = "^2.159.0"
google-auth-oauthlib = "^1.2.1"
tiktoken = "^0.8.0"
tree-sitter = "^0.24.0"
tree-sitter-c = "^0.23.4"
tree-sitter-java = "^0.23.5"
pydantic = "^2.10.5"
jinja2 = "^3.1.5"
pygithub = "^2.5.0"
langchain-experimental = "^0.3.4"
unidiff = "^0.7.5"
textdistance = "^4.6.3"
langchain-chroma = "^0.2.0"
ghapi = "^1.0.6"
opentelemetry-exporter-otlp-proto-grpc = "^1.29.0"
langchain-milvus = "^0.1.8"
xmltodict = "^0.14.2"
networkx = "^3.4.2"
matplotlib = "^3.10.0"
pydot = "^3.0.4"
art = "^6.4"
isort = "^5.13.2"
black = "^24.10.0"
deepdiff = "^8.1.1"
dicttoxml = "^1.7.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
